[Refacotr] - Membership : Moves membership classes to membership module
