+++ info/src/main/res/values/strings.xml
@@ -19,23 +19,11 @@
     <string name="museumGoogleMapSearchQuery" translatable="false"><![CDATA[https://www.google.com/maps/search/?api=1&query=The+Art+Institute+of+Chicago]]></string>
     <string name="versionInfo">Version %1$s</string>
 
-    <string name="signIn">Sign In</string>
-    <string name="homeZipCode">Zip Code</string>
-    <string name="memberId">Member ID</string>
-    <string name="changeInformation">Change Information</string>
-    <string name="switchCardHolder">Switch Cardholder</string>
-    <string name="memberCard">Member Card</string>
-    <string name="errorDialogTitle">Error</string>
-    <string name="ok" translatable="false">@android:string/ok</string>
     <string name="informationBannerSubtitle">Access your member card, see museum hours, or update your app settings.</string>
     <string name="alreadyMember">Already a Member?</string>
     <string name="whereAreYouLabel">Where are you?</string>
     <string name="languageSettingsBody">Some content may not be available in your selected language.</string>
-    <string name="english" translatable="false">English</string>
-    <string name="spanish" translatable="false">Español</string>
-    <string name="chinese" translatable="false">中文</string>
     <string name="chooseYourLanguage">Please Choose Your Preferred Language</string>
-    <string name="accessMemberCardLabel">Access Member Card</string>
     <string name="fragmentInformationLabel" translatable="false">fragment_information</string>
     <string name="information">Information</string>
     <string name="expires">Expires %s</string>
