+++ membership/src/main/java/edu/artic/membership/MemberInfoResponse.kt
@@ -1,4 +1,4 @@
-package edu.artic.info
+package edu.artic.membership
 
 import org.simpleframework.xml.Element
 import org.simpleframework.xml.Namespace
