+++ membership/src/main/res/layout/layout_member_access_card.xml
