+++ membership/src/main/res/values-es/strings.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="signIn">Iniciar sesión</string>
+    <string name="homeZipCode">Código postal</string>
+    <string name="memberId">Identificación del Miembro…</string>
+    <string name="changeInformation">Cambiar información</string>
+    <string name="switchCardHolder">Cambiar titular de tarjeta</string>
+    <string name="memberCard">Tarjeta del Miembro</string>
+    <string name="accessMemberCardLabel">Tarjeta de acceso para miembros</string>
+</resources>
