+++ membership/src/main/res/values/strings.xml
@@ -1,0 +1,15 @@
+<resources>
+    <string name="english" translatable="false">English</string>
+    <string name="spanish" translatable="false">Español</string>
+    <string name="chinese" translatable="false">中文</string>
+
+    <string name="signIn">Sign In</string>
+    <string name="homeZipCode">Zip Code</string>
+    <string name="memberId">Member ID</string>
+    <string name="changeInformation">Change Information</string>
+    <string name="switchCardHolder">Switch Cardholder</string>
+    <string name="memberCard">Member Card</string>
+    <string name="errorDialogTitle">Error</string>
+    <string name="ok" translatable="false">@android:string/ok</string>
+    <string name="accessMemberCardLabel">Access Member Card</string>
+</resources>
